feat: moves pixel helpers into modules folder
